requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
torch
torchvision
Pillow
numpy
opencv-python

